+++ src/context.ts
@@ -1,3 +1,4 @@
+import {Assertion} from './assertion'
 import {Capture} from './capture'
 import Environment from './environment'
 import Param from './param'
@@ -7,4 +8,5 @@ export default interface Context {
   environment: Environment;
   params: Param[];
   captures: Capture[];
+  assertions: Assertion[];
 }